fix: move UseCasesConfig from config to config/bean
