+++ src/main/java/br/com/fiap/tech_challenge/config/bean/DeleteProductByIdUseCaseConfig.java
@@ -1,4 +1,4 @@
-package br.com.fiap.tech_challenge.config;
+package br.com.fiap.tech_challenge.config.bean;
 
 import br.com.fiap.tech_challenge.core.domain.ports.ProductPersistence;
 import br.com.fiap.tech_challenge.core.domain.usecases.product.impl.DeleteProductByIdUseCaseImpl;
